+++ mapmemory-coroutines/src/main/kotlin/CoroutinesAccessors.kt
@@ -1,17 +1,15 @@
 package com.redmadrobot.mapmemory
 
-import com.redmadrobot.mapmemory.internal.getOrPutProperty
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlin.properties.ReadWriteProperty
 
 /**
  * Creates a delegate for dealing with [MutableStateFlow] stored in [MapMemory].
  * The delegate returns (and stores) new StateFlow if there is no corresponding value in memory.
  */
-public fun <T> MapMemory.stateFlow(initialValue: T): ReadWriteProperty<Any?, MutableStateFlow<T>> {
-    return getOrPutProperty { MutableStateFlow(initialValue) }
+public fun <T> MapMemory.stateFlow(initialValue: T): MapMemoryProperty<MutableStateFlow<T>> {
+    return invoke { MutableStateFlow(initialValue) }
 }
 
 /**
@@ -22,6 +20,6 @@ public fun <T> MapMemory.sharedFlow(
     replay: Int = 0,
     extraBufferCapacity: Int = 0,
     onBufferOverflow: BufferOverflow = BufferOverflow.SUSPEND,
-): ReadWriteProperty<Any?, MutableSharedFlow<T>> {
-    return getOrPutProperty { MutableSharedFlow(replay, extraBufferCapacity, onBufferOverflow) }
+): MapMemoryProperty<MutableSharedFlow<T>> {
+    return invoke { MutableSharedFlow(replay, extraBufferCapacity, onBufferOverflow) }
 }